Initial commit
I'm trying to do this with TypeScript but having a lot of bad luck.
Mainly the issue is that we don't want to emit modules, but in order to
run tests (which run in Node-land) we kind of need a way to load
modules. Jest doesn't work and neither does Mocha. I'm starting to think
that the apps-script-oauth2 project does it right — they just use the
gas-local package to load their files from their test files.