Use Rollup instead of Webpack
The ultimate goal of this project is to produce ES3-compatible code that
can be copied and pasted into Google's Script Editor (while still
allowing us to write the code in ES6). Because of this, we are ultra
concerned about the compiled version being 1) small and 2) human
readable. In that vein:

* Rollup produces more human-readable output than Webpack, which spits a
  glorified UMD preamble out before your actual code. So that's a win.
  However...
* I don't think we can use Lodash or Underscore. There's way too much
  code that both libraries produce. I tried to remove this by using
  babel-plugin-lodash, webpack-plugin-lodash, rollup-plugin-lodash, and
  even tools like Closure Compiler and Terser, since they have dead code
  elimination, but it help things much. So we're going to have to copy
  some of its features.
* If we ever want use advanced ES6 features such as Set that would
  require a polyfill in an ES3/5 environment, I don't think we can do
  that either, for the same reason.

And so. I think the only way to really accomplish this is be really
careful about what we write and to constantly monitor the compiled
version to make sure it doesn't get too big.